added CountryService and enhanced UX for country selection
