+++ CKODemoShop/ClientApp/src/app/interfaces/country.interface.ts
@@ -1,0 +1,5 @@
+export interface ICountry {
+  alpha2Code: string;
+  name: string;
+  flag: string;
+}
