+++ CKODemoShop/ClientApp/tsconfig.json
@@ -15,6 +15,7 @@
     ],
     "lib": [
       "es2018",
+      "esnext",
       "dom"
     ]
   }
